feat(weeklyContest): add reportCount to post response DTOs
  신고 기능을 위한 reportCount 필드 추가 및 PostInfo DTO 공통화
